Move PointService
